LayoutTests/css3/flexbox/css-properties-expected.txt

Tests being able to set the display to -webkit-flex and -webkit-inline-flex.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS flexbox.style.display is "-webkit-flex"
PASS flexbox.style.display is "inline"
PASS flexbox.style.display is "-webkit-inline-flex"
PASS flexbox.style.display is "-webkit-inline-flex"
PASS flexbox.style.display is "block"
PASS flexitem.style.webkitOrder is ""
PASS flexitem.style.webkitOrder is "2"
PASS flexitem.style.webkitOrder is "-1"
PASS flexitem.style.webkitOrder is "0"
PASS flexitem.style.webkitOrder is "0"
PASS flexitem.style.webkitOrder is "0"
PASS flexitem.style.webkitOrder is ""
PASS flexbox.style.webkitJustifyContent is ""
PASS window.getComputedStyle(flexbox, null).webkitJustifyContent is "normal"
PASS flexbox.style.webkitJustifyContent is ""
PASS flexbox.style.webkitJustifyContent is "flex-start"
PASS window.getComputedStyle(flexbox, null).webkitJustifyContent is "flex-start"
PASS flexbox.style.webkitJustifyContent is "flex-end"
PASS window.getComputedStyle(flexbox, null).webkitJustifyContent is "flex-end"
PASS flexbox.style.webkitJustifyContent is "center"
PASS window.getComputedStyle(flexbox, null).webkitJustifyContent is "center"
PASS flexbox.style.webkitJustifyContent is "space-between"
PASS window.getComputedStyle(flexbox, null).webkitJustifyContent is "space-between"
PASS flexbox.style.webkitJustifyContent is ""
PASS window.getComputedStyle(flexbox, null).webkitJustifyContent is "normal"
PASS flexbox.style.webkitAlignSelf is ""
PASS window.getComputedStyle(flexbox, null).webkitAlignSelf is "start"
PASS window.getComputedStyle(document.documentElement, null).webkitAlignSelf is "start"
PASS flexbox.style.webkitAlignSelf is ""
PASS window.getComputedStyle(flexbox, null).webkitAlignSelf is "start"
PASS flexbox.style.webkitAlignSelf is "auto"
PASS window.getComputedStyle(flexbox, null).webkitAlignSelf is "start"
PASS flexbox.style.webkitAlignSelf is "flex-start"
PASS window.getComputedStyle(flexbox, null).webkitAlignSelf is "flex-start"
PASS flexbox.style.webkitAlignSelf is "flex-end"
PASS window.getComputedStyle(flexbox, null).webkitAlignSelf is "flex-end"
PASS flexbox.style.webkitAlignSelf is "center"
PASS window.getComputedStyle(flexbox, null).webkitAlignSelf is "center"
PASS flexbox.style.webkitAlignSelf is "stretch"
PASS window.getComputedStyle(flexbox, null).webkitAlignSelf is "stretch"
PASS flexbox.style.webkitAlignSelf is "baseline"
PASS window.getComputedStyle(flexbox, null).webkitAlignSelf is "baseline"
PASS flexbox.style.webkitAlignSelf is ""
PASS window.getComputedStyle(flexbox, null).webkitAlignSelf is "start"
PASS flexbox.style.webkitAlignItems is ""
PASS flexitem.style.webkitAlignSelf is ""
PASS window.getComputedStyle(flexbox, null).webkitAlignItems is "stretch"
PASS window.getComputedStyle(flexitem, null).webkitAlignSelf is "stretch"
PASS flexbox.style.webkitAlignItems is ""
PASS flexitem.style.webkitAlignSelf is ""
PASS window.getComputedStyle(flexbox, null).webkitAlignItems is "stretch"
PASS window.getComputedStyle(flexitem, null).webkitAlignSelf is "stretch"
PASS flexbox.style.webkitAlignItems is "auto"
PASS flexitem.style.webkitAlignSelf is ""
PASS window.getComputedStyle(flexbox, null).webkitAlignItems is "stretch"
PASS window.getComputedStyle(flexitem, null).webkitAlignSelf is "stretch"
PASS flexbox.style.webkitAlignItems is "flex-start"
PASS flexitem.style.webkitAlignSelf is ""
PASS window.getComputedStyle(flexbox, null).webkitAlignItems is "flex-start"
PASS window.getComputedStyle(flexitem, null).webkitAlignSelf is "flex-start"
PASS flexbox.style.webkitAlignItems is "flex-end"
PASS window.getComputedStyle(flexbox, null).webkitAlignItems is "flex-end"
PASS window.getComputedStyle(flexitem, null).webkitAlignSelf is "flex-end"
PASS flexbox.style.webkitAlignItems is "center"
PASS window.getComputedStyle(flexbox, null).webkitAlignItems is "center"
PASS window.getComputedStyle(flexitem, null).webkitAlignSelf is "center"
PASS flexbox.style.webkitAlignItems is "stretch"
PASS window.getComputedStyle(flexbox, null).webkitAlignItems is "stretch"
PASS window.getComputedStyle(flexitem, null).webkitAlignSelf is "stretch"
PASS flexbox.style.webkitAlignItems is "baseline"
PASS window.getComputedStyle(flexbox, null).webkitAlignItems is "baseline"
PASS window.getComputedStyle(flexitem, null).webkitAlignSelf is "baseline"
PASS flexbox.style.webkitAlignItems is ""
PASS window.getComputedStyle(flexbox, null).webkitAlignItems is "stretch"
PASS window.getComputedStyle(flexitem, null).webkitAlignSelf is "stretch"
PASS flexbox.style.webkitAlignItems is ""
PASS window.getComputedStyle(flexbox, null).webkitAlignItems is "auto"
PASS window.getComputedStyle(flexitem, null).webkitAlignSelf is "auto"
PASS window.getComputedStyle(detachedFlexbox, null).webkitAlignSelf is ""
PASS window.getComputedStyle(detachedFlexItem, null).webkitAlignSelf is ""
PASS flexbox.style.webkitFlexDirection is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexDirection is "row"
PASS flexbox.style.webkitFlexDirection is ""
PASS flexbox.style.webkitFlexDirection is "row"
PASS window.getComputedStyle(flexbox, null).webkitFlexDirection is "row"
PASS flexbox.style.webkitFlexDirection is "row-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexDirection is "row-reverse"
PASS flexbox.style.webkitFlexDirection is "column"
PASS window.getComputedStyle(flexbox, null).webkitFlexDirection is "column"
PASS flexbox.style.webkitFlexDirection is "column-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexDirection is "column-reverse"
PASS flexbox.style.webkitFlexWrap is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexWrap is "nowrap"
PASS flexbox.style.webkitFlexWrap is ""
PASS flexbox.style.webkitFlexWrap is "nowrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexWrap is "nowrap"
PASS flexbox.style.webkitFlexWrap is "wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexWrap is "wrap"
PASS flexbox.style.webkitFlexWrap is "wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexWrap is "wrap-reverse"
PASS flexbox.style.webkitFlexFlow is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row nowrap"
PASS flexbox.style.webkitFlexFlow is ""
PASS flexbox.style.webkitFlexFlow is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row nowrap"
PASS flexbox.style.webkitFlexFlow is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row nowrap"
PASS flexbox.style.webkitFlexFlow is "nowrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row nowrap"
PASS flexbox.style.webkitFlexFlow is "nowrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row nowrap"
PASS flexbox.style.webkitFlexFlow is "wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row wrap"
PASS flexbox.style.webkitFlexFlow is "wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row wrap"
PASS flexbox.style.webkitFlexFlow is "wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row wrap-reverse"
PASS flexbox.style.webkitFlexFlow is "wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row wrap-reverse"
PASS flexbox.style.webkitFlexFlow is "row"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row nowrap"
PASS flexbox.style.webkitFlexFlow is "row"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row nowrap"
PASS flexbox.style.webkitFlexFlow is "row nowrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row nowrap"
PASS flexbox.style.webkitFlexFlow is "row nowrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row nowrap"
PASS flexbox.style.webkitFlexFlow is "row wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row wrap"
PASS flexbox.style.webkitFlexFlow is "row wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row wrap"
PASS flexbox.style.webkitFlexFlow is "row wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row wrap-reverse"
PASS flexbox.style.webkitFlexFlow is "row wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row wrap-reverse"
PASS flexbox.style.webkitFlexFlow is "row-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row-reverse nowrap"
PASS flexbox.style.webkitFlexFlow is "row-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row-reverse nowrap"
PASS flexbox.style.webkitFlexFlow is "row-reverse nowrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row-reverse nowrap"
PASS flexbox.style.webkitFlexFlow is "row-reverse nowrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row-reverse nowrap"
PASS flexbox.style.webkitFlexFlow is "row-reverse wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row-reverse wrap"
PASS flexbox.style.webkitFlexFlow is "row-reverse wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row-reverse wrap"
PASS flexbox.style.webkitFlexFlow is "row-reverse wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row-reverse wrap-reverse"
PASS flexbox.style.webkitFlexFlow is "row-reverse wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row-reverse wrap-reverse"
PASS flexbox.style.webkitFlexFlow is "column"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column nowrap"
PASS flexbox.style.webkitFlexFlow is "column"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column nowrap"
PASS flexbox.style.webkitFlexFlow is "column nowrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column nowrap"
PASS flexbox.style.webkitFlexFlow is "column nowrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column nowrap"
PASS flexbox.style.webkitFlexFlow is "column wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column wrap"
PASS flexbox.style.webkitFlexFlow is "column wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column wrap"
PASS flexbox.style.webkitFlexFlow is "column wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column wrap-reverse"
PASS flexbox.style.webkitFlexFlow is "column wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column wrap-reverse"
PASS flexbox.style.webkitFlexFlow is "column-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column-reverse nowrap"
PASS flexbox.style.webkitFlexFlow is "column-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column-reverse nowrap"
PASS flexbox.style.webkitFlexFlow is "column-reverse nowrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column-reverse nowrap"
PASS flexbox.style.webkitFlexFlow is "column-reverse nowrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column-reverse nowrap"
PASS flexbox.style.webkitFlexFlow is "column-reverse wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column-reverse wrap"
PASS flexbox.style.webkitFlexFlow is "column-reverse wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column-reverse wrap"
PASS flexbox.style.webkitFlexFlow is "column-reverse wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column-reverse wrap-reverse"
PASS flexbox.style.webkitFlexFlow is "column-reverse wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column-reverse wrap-reverse"
PASS flexbox.style.webkitFlexFlow is ""
PASS flexbox.style.webkitFlexFlow is ""
PASS flexbox.style.webkitFlexFlow is ""
PASS flexbox.style.webkitFlexFlow is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row nowrap"
PASS flexbox.style.webkitFlexFlow is "column"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column nowrap"
PASS flexbox.style.webkitFlexFlow is "column wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column wrap"
PASS flexbox.style.webkitFlexFlow is "row-reverse wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row-reverse wrap-reverse"
PASS foundFlexFlow is false
PASS flexbox.style.webkitAlignContent is ""
PASS window.getComputedStyle(flexbox, null).webkitAlignContent is "normal"
PASS flexbox.style.webkitAlignContent is "flex-start"
PASS window.getComputedStyle(flexbox, null).webkitAlignContent is "flex-start"
PASS flexbox.style.webkitAlignContent is "flex-end"
PASS window.getComputedStyle(flexbox, null).webkitAlignContent is "flex-end"
PASS flexbox.style.webkitAlignContent is "center"
PASS window.getComputedStyle(flexbox, null).webkitAlignContent is "center"
PASS flexbox.style.webkitAlignContent is "space-between"
PASS window.getComputedStyle(flexbox, null).webkitAlignContent is "space-between"
PASS flexbox.style.webkitAlignContent is "space-around"
PASS window.getComputedStyle(flexbox, null).webkitAlignContent is "space-around"
PASS flexbox.style.webkitAlignContent is "stretch"
PASS window.getComputedStyle(flexbox, null).webkitAlignContent is "stretch"
PASS flexbox.style.webkitAlignContent is ""
PASS window.getComputedStyle(flexbox, null).webkitAlignContent is "normal"
PASS flexbox.style.webkitAlignContent is ""
PASS window.getComputedStyle(flexbox, null).webkitAlignContent is "normal"
PASS successfullyParsed is true

TEST COMPLETE
